refactor a bit to make sure the fake clone/get are shared between packages
